+++ documentation/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 19.2.0
 
 * New: Added support for the E-mu Emulator IV bank format (E4B). Written banks have not been tested on real hardware yet.
+* New: E-mu Emulator IV banks can also be read directly from CD-ROM and hard disk images of the EOS samplers (ISO, IMG, HDA), including via the ISO/IMG source format.
 
 ## 19.1.0